+++ src/app/shared/shared.module.ts
@@ -2,17 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { BgHighlighterDirective } from './directives/bg-highlighter.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 
 @NgModule({
   declarations: [
-    BgHighlighterDirective
+    BgHighlighterDirective,
+    OrderByPipe
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    BgHighlighterDirective
+    BgHighlighterDirective,
+    OrderByPipe
   ]
 })
 export class SharedModule { }